[bw_assets] Add MPQ Module
MPQ is the archive format that Blizzard uses to store all its asset files. Because i am using the Amethyst game engine, I need to be able to:

- Parse the mpq archives
- Read files from the mpq archives concurrently
- Treat the archive as a "source" as if it was a directory

This change achieves this by wrapping the ceres_mpq archive in a new type that implements a read_file as a trait. It also provides an MPQSource struct that aggregates multiple MPQ files as a single source.